+++ src/error.rs
@@ -1,6 +1,5 @@
 use std::error::Error as StdError;
 use std::fmt;
-use std::str::Utf8Error;
 
 #[derive(Debug)]
 /// An error that occurred while performing an Atom operation.
@@ -8,8 +7,6 @@ use std::str::Utf8Error;
 pub enum Error {
     /// Unable to parse XML.
     Xml(XmlError),
-    /// Unable to parse UTF8 in to a string.
-    Utf8(Utf8Error),
     /// Input did not begin with an opening feed tag.
     InvalidStartTag,
     /// Unexpected end of input.
@@ -29,7 +26,6 @@ impl StdError for Error {
     fn source(&self) -> Option<&(dyn StdError + 'static)> {
         match *self {
             Error::Xml(ref err) => Some(err),
-            Error::Utf8(ref err) => Some(err),
             Error::InvalidStartTag => None,
             Error::Eof => None,
             Error::WrongDatetime(_) => None,
@@ -42,7 +38,6 @@ impl fmt::Display for Error {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match *self {
             Error::Xml(ref err) => fmt::Display::fmt(err, f),
-            Error::Utf8(ref err) => fmt::Display::fmt(err, f),
             Error::InvalidStartTag => write!(f, "input did not begin with an opening feed tag"),
             Error::Eof => write!(f, "unexpected end of input"),
             Error::WrongDatetime(ref datetime) => write!(
@@ -65,12 +60,6 @@ impl fmt::Display for Error {
 impl From<XmlError> for Error {
     fn from(err: XmlError) -> Error {
         Error::Xml(err)
-    }
-}
-
-impl From<Utf8Error> for Error {
-    fn from(err: Utf8Error) -> Error {
-        Error::Utf8(err)
     }
 }